Prerender: Disable prefetch if there's an appcache (with nonflakey tests)
Stops the preload scanner if an appcache manifest is detected and the document
is prefetching.

This un-reverts cl/2790323003. That test introduced by that CL was flakey due to
a race between AppCache creation and prefetching. This CL fixes the race by
blocking until the appropriate AppCache can be verified as existing.

BUG=632368

Review-Url: https://codereview.chromium.org/2819523002
Cr-Original-Commit-Position: refs/heads/master@{#464713}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f0abffcea31f665a2b1cc1979a955cf308c4dcb4