Fix backface-visibility with will-change
In the case where there is a scrollbar with will-change property and its
backface-visibility is set to hidden, the scrollbar will show up when
it is back-flipped. The reason is that we are not setting the backface
visibility for the overflow_controls_host_layer_.

A layout test is added.

BUG=700013
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:linux_layout_tests_slimming_paint_v2

Review-Url: https://codereview.chromium.org/2818533004
Cr-Original-Commit-Position: refs/heads/master@{#464782}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ab9fb6945b8d58a04bf92eb4dd16f0cd585da359