Remove Mac10.9 expectation for text-match-document-change-expected.png
The png for virtual/disable-spinvalidation/ appears to be broken anyway, and suspected to be causing
failures on
https://build.chromium.org/p/chromium.webkit/builders/WebKit%20Mac10.9/.

A Mac image expectation was added in
https://codereview.chromium.org/2824653002 and should cover Mac10.9
once this is removed.

Also remove the non-virtual mac-mac10.9 png, as it appears to be the same as the more generic Mac png from https://codereview.chromium.org/2824653002.

BUG=711689
TBR=hs1217.lee@samsung.com
NOTRY=true

Review-Url: https://codereview.chromium.org/2827453002
Cr-Original-Commit-Position: refs/heads/master@{#464932}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0fe1451d08ef167436488e7dbacbd8d840a5c858