[DevTools] Swallow mouse up in inspect mode
Currently, we dispatch mouseup to the page after selecting an element in inspect
mode. This could trigger any incovenient side-effects including removal of
inspected element. Swallowing mouseup fixes that problem.

This is not a 100% solution: overlay disappears if user holds mouse
for 2 seconds, and mouseup still gets to the page. That's a minor problem, so
this solution is good enough.

BUG=711320

Review-Url: https://codereview.chromium.org/2818883002
Cr-Original-Commit-Position: refs/heads/master@{#464986}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 88ebc67e440fa500369a453b32ac5f32a7dcbc64