Refactor DocumentMarkerController on top of new DocumentMarkerListEditor class
Based on yosin's comment on a previous version of this CL:
https://codereview.chromium.org/2812423002#msg7

This CL is step 1, "Introduce DMListEditor with in-place changes of DMC: Except for DMListEditor.h, changes are small". In this CL, I'm pulling out some of the
DocumentMarkerController implementation code into a DocumentMarkerListEditor class.
In a future CL, I will move this class into its own .h/.cpp files (I'm leaving it
in place for now to make the patch smaller).

BUG=707867

Review-Url: https://codereview.chromium.org/2812423002
Cr-Original-Commit-Position: refs/heads/master@{#465092}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b59bf346862b89daee2a4f06f02f5d7df59f9164