Worker: Raise an exception when an invalid feature number is passed to WorkerInternals::CountFeature
Before this CL, when an invalid feature number[1] is passed to
WorkerInternals::CountFeature/CountDeprecation, it ends up a check failure. This
is not a problem in real world applications because the internals APIs are used
only for testing, but it makes the fuzzer sad (see the report in the issue).

To make the fuzzer happy, this CL makes the functions raise an exception when
such an invalid value is given.

[1] feature < 0 || UseCounter::kNumberOfFeatures < feature

BUG=712003

Review-Url: https://codereview.chromium.org/2820193003
Cr-Original-Commit-Position: refs/heads/master@{#465137}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4b267ce12fab17dfa38615933571eae96faf63f1