[DevTools] Turn event listener breakpoints setting into session setting
Currently the setting is persisted (locally), but it is actually not applied
until the user manually opens up Event Listener Breakpoints pane.
This behavior is confusing, and we better make this setting not persist at all.

The alternative is to apply event listener breakpoints on frontend load,
but then it's easy to forget about such a breakpoint being set and it's hard
to connect sudden break with a place in UI to disable it.

BUG=none

Review-Url: https://codereview.chromium.org/2820263003
Cr-Original-Commit-Position: refs/heads/master@{#465349}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 484866d443789a5e284ddec3cf997430ffed4581