Rebaseline css2.1/t040304-c64-uri-00-a-g.html
Rebaseline failing layout test and remove it from TestExpectations.
From manually inspecting the test results, it seems the difference
is in the font of the output text, and not the core of the
expectation, that the three backgrounds of the text are green.

BUG=685951

Review-Url: https://codereview.chromium.org/2809593002
Cr-Original-Commit-Position: refs/heads/master@{#465482}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8d60ce0f354a569aaef18dd9547171a483b31525