Replicate feature policy container policies.
This CL adds a persistent parsed container policy to
HTMLFrameOwnerElement, and replicates it alongside of sandbox attributes
to remote frames. In most cases, the sandbox and FP attributes should be
processed at the same time, so the IPC and methods for them have been
combined. The combination of sandbox and FP is referred to as the "frame
policies".

BUG=682258
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:linux_site_isolation

Review-Url: https://codereview.chromium.org/2797813002
Cr-Original-Commit-Position: refs/heads/master@{#465563}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 92f8c0b3ff53422eb761b53793bac0b424ec45df