Remove PaintCanvas::writePixels
This method doesn't make a lot of sense when building a recording, and
it is not used for recordings for that reason. This changes callers of
it to call it on the SkSurface's SkCanvas directly instead by moving the
calls into Un/AcceleratedImageBufferSurface.

BUG=671433
CQ_INCLUDE_TRYBOTS=master.tryserver.blink:linux_trusty_blink_rel

Review-Url: https://codereview.chromium.org/2822263004
Cr-Original-Commit-Position: refs/heads/master@{#465805}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6527e1ef5c38b22789f94f5ae4c763031f5932f7