Hide cold mode idle time spellchecker behind a flag
This patch adds a new runtime flag IdleTimeColdModeSpellChecking,
so that idle time spellchecker enters cold mode only when the flag
is enabled --- otherwise it enters the inactive state instead.

The unit tests are modified accordingly so that the lifecycle transition
without the flag is also tested.

The purpose of the new flag is for shipping hot and cold modes of
idle time spellchecker separately.

BUG=713006

Review-Url: https://codereview.chromium.org/2827903002
Cr-Original-Commit-Position: refs/heads/master@{#465855}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 234a68c81bb610d29c3bcabc08dcd4485ee58f06