Don't decode move setter on external fields in ComputedStyle.
External fields such as 'width' have non bitfield types such as Length.
We generate move setters for these fields that take its argument by
rvalue reference:

void SetWidth(Length&& v) {
  width_ = std::move(v);
}

However, in the Jinja template, we use the Jinja macro decode on the
value v. The decode macro tries to perform a static_cast if the field
is a bitfield. But since external fields can never be a bitfield, this
is a no-op.

This patch removes the call to the decode macro.

BUG=628043

Review-Url: https://codereview.chromium.org/2830793003
Cr-Original-Commit-Position: refs/heads/master@{#465876}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ecbdd0c68e34edbb28283db17f52cb3fa7f95afc