Messaging: Avoid calling non-thread-safe functions in MessagePort::messageAvailable()
As implementation comments in MessagePort::messageAvailable(), it can be called
from another thread and should not call non-thread-functions. However, the
current implementation wrongly calls ContextLifecycleObserver's
GetExecutionContext() that is not a thread-safe function to post a task to the
context thread.

To avoid that, this CL replaces ExecutionContext::PostTask() with
WebTaskRunner::PostTask(). The task runner is captured in the ctor of
MessagePort called on the context thread.

BUG=694925

Review-Url: https://codereview.chromium.org/2818073002
Cr-Original-Commit-Position: refs/heads/master@{#465886}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 829907ac0d3d2ed39d8752d22a6d0e45df1e8325