Remove 'name' parameter from CSSPropertyAPIMethods.json5.
In the parameters section of CSSPropertyAPIMethods.json5, there is an
entry for 'name'. However, 'name' is a special parameter that is
used by the json5 reader. In this case, the 'name' entry is superfluous.
This patch removes the 'name' entry and adds an assert to prevent
future uses of 'name'.

BUG=628043

Review-Url: https://codereview.chromium.org/2806163002
Cr-Original-Commit-Position: refs/heads/master@{#465891}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c695cc71ebc3348aa9db521f7f4324502a0134e3