[css-grid] Remove most of the usage of SizingOperation
SizingOperation was added as a way to modify the behaviour of the track
sizing algorithm with the specifics for intrinsic size computations. The
problem is that although intrinsic size computation does imply indefinite
sizes, the opposite is not always true. For example, a grid container with
height:auto would compute the row sizes with an  indefinite size but is not
part of the intrinsic size (preferred widths) computation.

That's why it was wrongly used in some parts of the current code, in most of
the cases is more than enough to check whether the available space is
indefinite or not. In order to do that we have to convert available size
into an Optional to properly represent indefinite sizes. Also we should keep
both available sizes (height & width) around all the time to properly
support orthogonal flows.

Review-Url: https://codereview.chromium.org/2823183003
Cr-Original-Commit-Position: refs/heads/master@{#466019}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: db336d8fa4c9af65243356ea712122d6f8e15cf1