Media Controls: move timeline related code to modules.
This is moving the timeline slider code and its metrics helper class. It
is also moving a method out of MediaControls.h that should have happen
in a different CL. Because of headers being reduced, the CL also
requires to add a few headers to implement files.

BUG=662761
R=avayvod@chromium.org,jochen@chromium.org

Review-Url: https://codereview.chromium.org/2820343002
Cr-Original-Commit-Position: refs/heads/master@{#466048}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 322f0d1c2543907fcc64ebcf55845d91c8b7e88b