Hide popups when handling mouse down, mouse wheel, or gesture tap in a WebFrameWidget
When a mouse down is handled inside WebViewImpl (main frame), the current page
popup is hidden. The same behavior should apply to out of process renderers.
However, in OOPIF renderers the WebViewImpl does not handle mouse events and
the task is delegated to the WebFrameWidget corresponding to the frame which
has received the event. But since receiving and input inside a WebFrameWidget
suggests that the mouse down has been outside the bounds of the WebPagePopup,
we should dismiss any currently showing popups the same way as we do it for
WebViewImpls.

Specifically, not having this behavior led to a bug on Windows where date
picker inside OOPIFs does not show again after dismissing an older one by
clicking inside the OOPIF.

BUG=671732

Review-Url: https://codereview.chromium.org/2804813002
Cr-Original-Commit-Position: refs/heads/master@{#466055}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e3cbfeec943500d99832b93b015d62937ff15443