[refactor] Fix autofill features for payments when the form is inside an OOPIF
Currently, clicking into a credit card information <input> field inside an
OOPIF does not pop up the autofill window. The root cause is simply because
the call to RenderWidget::FocusChangeComplete() is dropped for OOPIFs.
Moreover, for the main frame, this call is forwarded throught
RenderViewObserver calls to the corresponding PageClickTracker.

On the other hand, both PageClickTracker and AutofillAgent use a internal
Legacy class to observe RenderView for updates in element focus and mouse down
down inside a node. This is unfortunate given that the actual classes are
RenderFrameObservers.

This CL will make the following changes:

A) FocusChangeComplete:
This call is now forwarded to the observers of the RenderWidget which are
RenderFrames. They will then notify their own observers, e.g., AutofillAgent
and PageClickTracker. In line with this, the Legacy classes inside these
classes will be removed. Also, there is no longer a need for
RenderViewImpl::RenderWidgetFocusChangeComplete.

B) MouseDown:
Currently, in response to a left mouse button down or a gesture tap, ChromeClient
notifies the WebViewClient (RenderViewImpl). This is solely used in
PageClickTracker. This CL will remove the call from WebViewClient to the
WebWidgetClient (RenderWidget) which similarly to FocusChangeComplete, will be
forwarded to the observer RenderFrames and eventually the RenderFrameObservers.

The change in A) will also fix a bug with credit card payment autofill in
OOPIFs.

BUG=712754,703800, 583347, 433486

Review-Url: https://codereview.chromium.org/2766053002
Cr-Original-Commit-Position: refs/heads/master@{#466078}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 27ca69b1e049f9f86e15f34420eeef224335f075