Increase the number of layout test smoke tests.
This CL adds more tests to LayoutTests/SmokeTests so that that
there are about 1060 tests now. The tests there a sampling
of the fastest tests from each directory.

The main effect of this is to increase the number of layout tests
run on android on the waterfall and try bots.

This CL also removes existing android baselines so that the set of
baselines in LayoutTests/platform/android are baselines for tests
that are actually run.

Bug: 412732
Change-Id: Ic0277fd90bba3153c558d883a046a75f62ee74bb
Reviewed-on: https://chromium-review.googlesource.com/481042
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#466406}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 049ab8b91cb873da45755e6c678d153b7fc3d856