Move most of FrameLoader::CheckCompleted() to Document
Split out the the parts that are specific to the committed Document, and
move those to a new Document::CheckCompleted(). Most current callers of
FrameLoader::CheckCompleted() will now call Document::CheckCompleted()
instead. Rename the remained of FrameLoader::CheckCompleted() to
DidFinishNavigation(), and have it handling the logic that is tied to
setting Frame::loading_ to false and firing DidStopLoading() callbacks.

BUG=

Review-Url: https://codereview.chromium.org/2809733003
Cr-Original-Commit-Position: refs/heads/master@{#466470}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 37861ef99134d28395b1891a25edb0b750a7011e